refactor: convert integration TestCRDValidations to envtest
